requirements.txt
yfinance==0.2.43
pillow==10.4.0
pandas==2.2.3
pytz==2024.1
